+++ tests/conftest.py
@@ -75,27 +75,27 @@ def api_url():
 
 
 @pytest.fixture
-def app_id():
-    return "fake-app-id"
+def client_id():
+    return "fake-client-id"
 
 
 @pytest.fixture
-def app_secret():
+def client_secret():
     return "nyl4n4ut"
 
 
 @pytest.fixture
 def api_client(api_url):
     return APIClient(
-        app_id=None, app_secret=None, access_token=None, api_server=api_url
+        client_id=None, client_secret=None, access_token=None, api_server=api_url
     )
 
 
 @pytest.fixture
-def api_client_with_app_id(access_token, api_url, app_id, app_secret):
+def api_client_with_client_id(access_token, api_url, client_id, client_secret):
     return APIClient(
-        app_id=app_id,
-        app_secret=app_secret,
+        client_id=client_id,
+        client_secret=client_secret,
         access_token=access_token,
         api_server=api_url,
     )
@@ -150,7 +150,7 @@ def mock_account(mocked_responses, api_url, account_id):
 
 
 @pytest.fixture
-def mock_accounts(mocked_responses, api_url, account_id, app_id):
+def mock_accounts(mocked_responses, api_url, account_id, client_id):
     accounts = [
         {
             "account_id": account_id,
@@ -173,7 +173,9 @@ def mock_accounts(mocked_responses, api_url, account_id, app_id):
             return (200, {}, json.dumps([]))
         return (200, {}, json.dumps(accounts))
 
-    url_re = "{base}(/a/{app_id})?/accounts/?".format(base=api_url, app_id=app_id)
+    url_re = "{base}(/a/{client_id})?/accounts/?".format(
+        base=api_url, client_id=client_id
+    )
     mocked_responses.add_callback(
         responses.GET,
         re.compile(url_re),
@@ -1296,7 +1298,7 @@ def mock_resources(mocked_responses, api_url):
 
 
 @pytest.fixture
-def mock_account_management(mocked_responses, api_url, account_id, app_id):
+def mock_account_management(mocked_responses, api_url, account_id, client_id):
     account = {
         "account_id": account_id,
         "email_address": "ben.bitdiddle1861@gmail.com",
@@ -1311,11 +1313,11 @@ def mock_account_management(mocked_responses, api_url, account_id, app_id):
     account["billing_state"] = "cancelled"
     cancelled_response = json.dumps(account)
 
-    upgrade_url = "{base}/a/{app_id}/accounts/{id}/upgrade".format(
-        base=api_url, id=account_id, app_id=app_id
+    upgrade_url = "{base}/a/{client_id}/accounts/{id}/upgrade".format(
+        base=api_url, id=account_id, client_id=client_id
     )
-    downgrade_url = "{base}/a/{app_id}/accounts/{id}/downgrade".format(
-        base=api_url, id=account_id, app_id=app_id
+    downgrade_url = "{base}/a/{client_id}/accounts/{id}/downgrade".format(
+        base=api_url, id=account_id, client_id=client_id
     )
     mocked_responses.add(
         responses.POST,
@@ -1334,9 +1336,9 @@ def mock_account_management(mocked_responses, api_url, account_id, app_id):
 
 
 @pytest.fixture
-def mock_revoke_all_tokens(mocked_responses, api_url, account_id, app_id):
-    revoke_all_url = "{base}/a/{app_id}/accounts/{id}/revoke-all".format(
-        base=api_url, id=account_id, app_id=app_id
+def mock_revoke_all_tokens(mocked_responses, api_url, account_id, client_id):
+    revoke_all_url = "{base}/a/{client_id}/accounts/{id}/revoke-all".format(
+        base=api_url, id=account_id, client_id=client_id
     )
     mocked_responses.add(
         responses.POST,
@@ -1348,9 +1350,9 @@ def mock_revoke_all_tokens(mocked_responses, api_url, account_id, app_id):
 
 
 @pytest.fixture
-def mock_ip_addresses(mocked_responses, api_url, app_id):
-    ip_addresses_url = "{base}/a/{app_id}/ip_addresses".format(
-        base=api_url, app_id=app_id
+def mock_ip_addresses(mocked_responses, api_url, client_id):
+    ip_addresses_url = "{base}/a/{client_id}/ip_addresses".format(
+        base=api_url, client_id=client_id
     )
     mocked_responses.add(
         responses.GET,
@@ -1372,9 +1374,9 @@ def mock_ip_addresses(mocked_responses, api_url, app_id):
 
 
 @pytest.fixture
-def mock_token_info(mocked_responses, api_url, account_id, app_id):
-    token_info_url = "{base}/a/{app_id}/accounts/{id}/token-info".format(
-        base=api_url, id=account_id, app_id=app_id
+def mock_token_info(mocked_responses, api_url, account_id, client_id):
+    token_info_url = "{base}/a/{client_id}/accounts/{id}/token-info".format(
+        base=api_url, id=account_id, client_id=client_id
     )
     mocked_responses.add(
         responses.POST,