+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased (dev)
 ----------------
 * Add `metadata` field in the Event model to support new event metadata feature
 * Add filtering support for `metadata_pair`
+* Fix issue when converting offset-aware `datetime` objects to `timestamp`
 
 v4.12.1
 -------