+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Unreleased (dev)
 * Add filtering support for `metadata_pair`
 * Fix adding a tracking object to an existing `draft`
 * Fix issue when converting offset-aware `datetime` objects to `timestamp`
+* Remove `bumpversion` from a required dependency to an extra dependency
 
 v4.12.1
 -------