+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 nylas-python Changelog
 ======================
 
-Unreleased (dev)
+v.5.0.0
 ----------------
 * Transitioned from `app_id` and `app_secret` naming to `client_id` and `client_secret`
 * Add support for the Nylas Neural API